+++ lib/active_type/record_extension/inheritance.rb
@@ -15,11 +15,11 @@ module ActiveType
           options = scope
           scope = nil
         end
-        unless options[:foreign_key]
+        unless options[:foreign_key] || options[:as]
           options = options.merge(foreign_key: extended_record_base_class.name.foreign_key)
         end
         if ActiveRecord::VERSION::MAJOR > 3
-          [scope, options]
+          [options, scope]
         else
           [options]
         end
@@ -28,19 +28,58 @@ module ActiveType
       module ClassMethods
 
         def model_name
-          extended_record_base_class.model_name
+          @_model_name ||= begin
+            if name
+              # Namespace detection copied from ActiveModel::Naming
+              namespace = module_ancestors.detect do |n|
+                n.respond_to?(:use_relative_model_naming?) && n.use_relative_model_naming?
+              end
+              # We create a Name object, with the sti class name, but self as the @klass reference
+              # This way lookup_ancestors is invoked on the right class instead of the extended_record_base_class
+              dup_model_name = ActiveModel::Name.new(self, namespace, sti_name)
+              key = name.underscore.to_sym
+              # We set the `i18n_key` to lookup on the derived class key
+              # We keep the others the same to preserve parameter and route names
+              dup_model_name.instance_variable_set(:@i18n_key, key)
+              dup_model_name
+            else # name is nil for the anonymous intermediate class
+              extended_record_base_class.model_name
+            end
+          end
+        end
+
+        def module_ancestors
+          if extended_record_base_class.respond_to?(:module_parents)
+            extended_record_base_class.module_parents
+          else
+            extended_record_base_class.parents
+          end
         end
 
         def sti_name
           extended_record_base_class.sti_name
         end
 
-        def has_many(name, *args, &extension)
-          super(name, *Inheritance.add_foreign_key_option(extended_record_base_class, *args), &extension)
+        def descends_from_active_record?
+          extended_record_base_class.descends_from_active_record?
         end
 
-        def has_one(name, *args, &extension)
-          super(name, *Inheritance.add_foreign_key_option(extended_record_base_class, *args), &extension)
+        def has_many(name, scope=nil, *args, &extension)
+          new_args, new_scope = Inheritance.add_foreign_key_option(extended_record_base_class, scope, *args)
+          if ActiveRecord::VERSION::MAJOR <= 3 || new_scope.nil?
+            super(name, **new_args, &extension)
+          else
+            super(name, new_scope, **new_args, &extension)
+          end
+        end
+
+        def has_one(name, scope=nil, *args, &extension)
+          new_args, new_scope = Inheritance.add_foreign_key_option(extended_record_base_class, scope, *args)
+          if ActiveRecord::VERSION::MAJOR <= 3 || new_scope.nil?
+            super(name, **new_args, &extension)
+          else
+            super(name, new_scope, **new_args, &extension)
+          end
         end
 
         private
@@ -49,7 +88,23 @@ module ActiveType
 
           def find_sti_class(type_name)
             sti_class = super
-            if self <= sti_class
+
+            # Consider this class hierarchy
+            # class Parent < ActiveRecord::Base; end
+            # class Child < Parent; end
+            # class ExtendedParent < ActiveType::Record[Parent]; end
+            # class ExtendedChild < ActiveType::Record[Child]; end
+            if self < sti_class
+              # i.e. ExtendendChild.find(child.id)
+              # => self = ExtendedChild; sti_class = Child
+              # instantiate as ExtendedChild
+              self
+            elsif sti_class < extended_record_base_class
+              # i.e. ExtendedParent.find(child.id)
+              # => sti_class = Child; self = ExtendedParent; extended_record_base_class = Parent
+              # There is no really good solution here, since we cannot instantiate as both ExtendedParent
+              # and Child. We opt to instantiate as ExtendedParent, since the other option can be
+              # achieved by using Parent.find(child.id)
               self
             else
               sti_class
@@ -77,7 +132,22 @@ module ActiveType
                 "or overwrite #{name}.inheritance_column to use another column for that information."
             end
             #### our code starts here
-            if self <= subclass
+            # Consider this class hierarchy
+            # class Parent < ActiveRecord::Base; end
+            # class Child < Parent; end
+            # class ExtendedParent < ActiveType::Record[Parent]; end
+            # class ExtendedChild < ActiveType::Record[Child]; end
+            if self < subclass
+              # i.e. ExtendendChild.find(child.id)
+              # => self = ExtendedChild; subclass = Child
+              # instantiate as ExtendedChild
+              subclass = self
+            elsif subclass < extended_record_base_class
+              # i.e. ExtendedParent.find(child.id)
+              # => subclass = Child; self = ExtendedParent; extended_record_base_class = Parent
+              # There is no really good solution here, since we cannot instantiate as both ExtendedParent
+              # and Child. We opt to instantiate as ExtendedParent, since the other option can be
+              # achieved by using Parent.find(child.id)
               subclass = self
             end
             #### our code ends here