+++ lib/active_type/util.rb
@@ -34,6 +34,11 @@ module ActiveType
         casted.instance_variable_set(:@new_record, record.new_record?)
         # Rails 3.2, 4.2
         casted.instance_variable_set(:@destroyed, record.destroyed?)
+        # Rails 5.2+
+        casted.instance_variable_set(:@mutations_from_database, record.instance_variable_get(:@mutations_from_database))
+
+        casted.instance_variable_set(:@association_cache, record.instance_variable_get(:@association_cache))
+
         # Rails 3.2, 4.2
         errors = record.errors
         if errors.kind_of? ActiveModel::Errors