+++ lib/active_type/virtual_attributes.rb
@@ -8,6 +8,23 @@ module ActiveType
 
   module VirtualAttributes
 
+    module Serialization
+      extend ActiveSupport::Concern
+
+      def init_with(coder)
+        if coder['virtual_attributes'].present?
+          @virtual_attributes = coder['virtual_attributes']
+        end
+        super(coder)
+      end
+
+      def encode_with(coder)
+        coder['virtual_attributes'] = @virtual_attributes
+        coder['active_type_yaml_version'] = 1
+        super(coder)
+      end
+    end
+
     class VirtualColumn
 
       def initialize(name, type_caster, options)
@@ -110,10 +127,10 @@ module ActiveType
       result
     end
 
-
     extend ActiveSupport::Concern
 
     included do
+      include ActiveType::VirtualAttributes::Serialization
       class_attribute :virtual_columns_hash
       self.virtual_columns_hash = {}
 
@@ -144,28 +161,53 @@ module ActiveType
       @virtual_attributes_cache ||= {}
     end
 
-    def [](name)
+    def read_existing_virtual_attribute(name, &block_when_not_virtual)
       if self.singleton_class._has_virtual_column?(name)
         read_virtual_attribute(name)
       else
-        super
+        yield
       end
     end
 
-    # ActiveRecord 4.2.1
-    def _read_attribute(name)
+    def write_existing_virtual_attribute(name, value, &block_when_not_virtual)
       if self.singleton_class._has_virtual_column?(name)
-        read_virtual_attribute(name)
+        write_virtual_attribute(name, value)
       else
-        super
+        yield
+      end
+    end
+
+    def [](name)
+      read_existing_virtual_attribute(name) { super }
+    end
+
+    if ActiveRecord::VERSION::STRING >= '4.2.0'
+      def _read_attribute(name)
+        read_existing_virtual_attribute(name) { super }
+      end
+    end
+
+    if ActiveRecord::VERSION::STRING < '4.2.0' || ActiveRecord::VERSION::STRING >= '6.1.0'
+      # in 6.1, read_attribute does not call _read_attribute
+      def read_attribute(name)
+        read_existing_virtual_attribute(name) { super }
       end
     end
 
     def []=(name, value)
-      if self.singleton_class._has_virtual_column?(name)
-        write_virtual_attribute(name, value)
-      else
-        super
+      write_existing_virtual_attribute(name, value) { super }
+    end
+
+    if ActiveRecord::VERSION::STRING >= '5.2.0'
+      def _write_attribute(name, value)
+        write_existing_virtual_attribute(name, value) { super }
+      end
+    end
+
+    if ActiveRecord::VERSION::STRING < '5.2.0' || ActiveRecord::VERSION::STRING >= '6.1.0'
+      # in 6.1, write_attribute does not call _write_attribute
+      def write_attribute(name, value)
+        write_existing_virtual_attribute(name, value) { super }
       end
     end